requirements.txt
flask
flask_sqlalchemy
slugify
flask_marshmallow
marshmallow_sqlalchemy

